[슝] Chapter 9. API & Paging

+++ src/main/java/com/example/spring/domain/review/dto/req/ReviewReqDTO.java
@@ -10,6 +10,6 @@ public class ReviewReqDTO {
             Long storeId,
             String comment,
             @NotNull
-            Integer rating
+            Long rating
     ){}
 }
